Speed up move files on windows
