+++ src-tauri/Cargo.toml
@@ -48,6 +48,9 @@ image = "0.25.6"
 rustls = { version = "0.23.31", default-features = false, features = ["ring"] }
 unxip-rs = "0.1.1"
 
+[target.'cfg(unix)'.dependencies]
+sdkmover = { path = "../sdkmover" }
+
 [features]
 # This feature is used for production builds or when a dev server is not specified, DO NOT REMOVE!!
 custom-protocol = ["tauri/custom-protocol"]
