+++ src-tauri/tauri.windows.conf.json
@@ -1,0 +1,5 @@
+{
+  "bundle": {
+    "resources": ["templates", "cpio", "sdkmoverbin"]
+  }
+}
